requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
passlib==1.7.4
python-decouple==3.8
django-cors-headers==4.3.1
drf-yasg==1.21.7